tests/spatnav-focus-001.html

<!doctype html>
<meta charset="utf-8">
<title>Spatnav sanity check</title>
  <link rel="author" title="Florian Rivoal" href="https://florian.rivoal.net">
  <meta name=flags content="">
  <link rel=help href="https://wicg.github.io/spatial-navigation/#high-level-api">

  <script src="../resources/testharness.js"></script>
  <script src="../resources/testharnessreport.js"></script>

<!--load polyfill. To be removed before submitting to WPT-->
<script src="../polyfill/spatnav-heuristic.js"></script>

<body>
  <table>
    <tr><td><button id=b1>b1</button><td><button id=b2>b2</button>
    <tr><td><button id=b3>b3</button><td><button id=b4>b4</button>
  </table>
</body>

<script>
test(function() {
  document.getElementById("b1").focus();
  window.navigate("down");
  assert_equals(document.activeElement.id, "b3");
  window.navigate("up");
  assert_equals(document.activeElement.id, "b1");
  window.navigate("right");
  assert_equals(document.activeElement.id, "b2");
  window.navigate("down");
  assert_equals(document.activeElement.id, "b4");
  window.navigate("left");
  assert_equals(document.activeElement.id, "b3");
}, "spatnav can move the focus in all 4 directions")

</script>
